Moved locationView to seperated view
